% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_label}
\alias{add_label}
\title{Adds useful labels to colData.
Can be used to add cell tyles to cluster numbers}
\usage{
add_label(fsce, labels, by = NULL, expt = "rnaseq")
}
\arguments{
\item{fsce}{An object of class \code{\link{FunctionalSingleCellExperiment}}}

\item{labels}{dataframe of new labels. Must contain at least one column of matching variables (e.g. cell_id or k_cluster)}

\item{by}{column name labels to match in colData of \code{expt}. If NULL, will match by all matching column names}

\item{expt}{Data to use for match labels
(default is \code{rnaseq}). Must be present in \code{names(fsce)}.}
}
\value{
fsce with all \code{labels} in \code{expt} colData.
}
\description{
Adds useful labels to colData.
Can be used to add cell tyles to cluster numbers
}
\examples{
# Add cell_type labels to PBMC data

labels <- data.frame(k_cluster = as.factor(c("1", "2", "3", "4", "5", "6")),
label = c("MC", "NK", "NK+T", "MC", "MK", "CD4/8 T"))

fsce <- add_label(fsce_small, labels)

SingleCellExperiment::colData(fsce[["rnaseq"]])

}
