% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add_stats}
\alias{add_stats}
\title{Add stats to plot activities per cluster}
\usage{
add_stats(p, df, val = q.value, y_loc = y_loc, xmin = group,
  xmax = group1)
}
\arguments{
\item{p}{plot to add stats to}

\item{df}{dataframe output from \code{\link[=stat_activity_grouped]{stat_activity_grouped()}}.
Output must include columnw with names group, group1, q.value and/or p.value.}

\item{val}{values to add to plot. Default is q.value from  \code{\link[=stat_activity_grouped]{stat_activity_grouped()}} output}

\item{y_loc}{location of comparisson bars on graph. Default it \code{y_loc} column of \code{df}. Can also be numeric values.}

\item{xmin}{start location of comparison bar. Default is \code{group} column of \code{df}.}

\item{xmax}{stop location of comparison bar. Default is \code{group1} column of \code{df}.}
}
\description{
Adds stat comparison bars to a beeswarm plot of activity across specified groups
}
\seealso{
Other plot fuctions: \code{\link{plot_heatmap}}
}
\concept{plot fuctions}
