% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{barnyard_data}
\alias{barnyard_data}
\title{Generate Seurat object from barnyard data}
\usage{
barnyard_data()
}
\description{
Load data from barnyard experiment peformed with
UNG and RNASEH2C Hap1 cells and A:U and rG hairpin
substrates.
}
