% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{calc_cell_cycle}
\alias{calc_cell_cycle}
\title{Calculate cell-cycle assignments}
\usage{
calc_cell_cycle(fsce, expt = "rnaseq", org = "human", ...)
}
\arguments{
\item{fsce}{\code{\link{FunctionalSingleCellExperiment}}}

\item{expt}{Data to use for calculating variable features
(default is \code{rnaseq}). Must be present in \code{names(fsce)}.}

\item{org}{organism for gene pairs, human or mouse}

\item{...}{additional params for \code{\link[scran:cyclone]{scran::cyclone()}}}
}
\description{
Uses \code{\link[scran:cyclone]{scran::cyclone()}} to assign cell-cycle phases to each cell.
Assignments are available in \code{colData} in the \code{cell_cycle} slot.
}
\examples{
\donttest{
fsce_small <- calc_cell_cycle(fsce_small)
SingleCellExperiment::colData(fsce_small[["rnaseq"]])
}

}
\references{
Scialdone A, Natarajana KN, Saraiva LR et al. (2015).
Computational assignment of cell-cycle stage from single-cell transcriptome data.
\emph{Methods} 85:54--61
}
