% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calc_group_stats}
\alias{calc_group_stats}
\title{Calculate statistics between pairs of groups}
\usage{
calc_group_stats(df, group)
}
\arguments{
\item{df}{tidied version of data from a \code{SingleCellExperiment}}

\item{group}{variable for generating combinations}
}
\description{
Applies \code{\link[stats:wilcox.test]{stats::wilcox.test()}} to all combinations of variables across
groups in a tidied data frame.
}
\details{
Data is assumed to contain a single grouping variable, and all other
variables are treated as activitys to compare. Descriptive columns like
\code{cell_id} should not be included.
}
\examples{
x <- select(fsce_tidy, k_cluster, Uracil_45, riboG_44)
x

calc_group_stats(x, group = k_cluster)

}
