% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{calc_kmeans}
\alias{calc_kmeans}
\title{K-means clustering}
\usage{
calc_kmeans(fsce, expt = "rnaseq", k, method = "PCA", n_dims = NULL,
  ...)
}
\arguments{
\item{fsce}{\code{\link{FunctionalSingleCellExperiment}}}

\item{expt}{Data to use for calculating variable features
(default is \code{rnaseq}). Must be present in \code{names(fsce)}.}

\item{k}{number of classes}

\item{method}{dimensionality reduction method for clustering (defaults to
PCA)}

\item{n_dims}{specify the number of dimensions from "dr" to use for
clustering, defaults to all dimensions}

\item{...}{additional arguments to pass to \code{\link[stats:kmeans]{stats::kmeans()}}}
}
\value{
fsce with \code{k_cluster} in \code{expt} colData.
}
\description{
K-means clustering
}
\examples{
# calculate PCA for k-means default method
fsce <- calc_pca(fsce_small)

fsce <- calc_kmeans(fsce, k = 6)
\dontrun{
library(SingleCellExperiment)
colData(fsce[["rnaseq"]], "k_cluster")
}

}
\concept{clustering}
