% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calc_qvalues}
\alias{calc_qvalues}
\title{Calculate q-values}
\usage{
calc_qvalues(x, id = "p.value", bh = TRUE)
}
\arguments{
\item{x}{data frame}

\item{id}{name of variable with p-values}

\item{bh}{If \code{TRUE}, use the Benjamini-Hochberg adjustment}
}
\description{
Calculate q-values
}
