% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{calc_signal_enrichment}
\alias{calc_signal_enrichment}
\title{Identify haircut signals enriched over empty droplet distribution}
\usage{
calc_signal_enrichment(fsce, unfiltered_mat, umi_cutoff = 50,
  count_cutoff = 10, qval_cutoff = 0.001, enrich_cutoff = 2,
  per_cluster = FALSE, cluster_type = "leiden_cluster")
}
\arguments{
\item{fsce}{An object of class \code{\link{FunctionalSingleCellExperiment}}}

\item{unfiltered_mat}{matrix with haircut umi counts for all barcodes, or path to
unfiltered matrix able to be read by \code{\link[scrunchy:read_matrix]{scrunchy::read_matrix()}}}

\item{umi_cutoff}{minimum umi counts used to define background. Droplets
with fewer than this number of UMIs are considered backgroud, default = 50}

\item{count_cutoff}{minimum UMI counts at each site required for testing, default = 10}

\item{qval_cutoff}{minimum q.value to consider as enriched,
default = 0.001}

\item{enrich_cutoff}{minimum log2 increase to consider as enriched,
default = 2}

\item{per_cluster}{Default = FALSE, calculate enrichment per cluster}

\item{cluster_type}{if per_cluster than clustering column to group by,
one of "k_cluster" or "leiden_cluster", defaults to "leiden_cluster".}
}
\value{
\code{fsce} with enrichment values added to haircut `rowData``
}
\description{
Identify haircut signals enriched over empty droplet distribution
}
