% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.R
\name{calc_umap}
\alias{calc_umap}
\title{Generate 2D cell embeddings using UMAP}
\usage{
calc_umap(fce, expt = "sce", dr = "PCA", n_dims = NULL,
  n_neighbors = 30, min_dist = 0.3, metric = "pearson2", seed = NA,
  ...)
}
\arguments{
\item{fce}{fce object}

\item{expt}{Data to use for UMAP, one of either sce (rna data, the default) or fsce (functional data)}

\item{dr}{select which dimenality reduction data to use for UMAP, defaults to PCA}

\item{n_dims}{number of dimensions to pass to UMAP, defaults to all present in dr matrix}

\item{n_neighbors}{number of nearest neighbors to use for learning the manifold. Low values will
preserve local structure, at the expense missing higher order organization. Higher values will
capture more global structure but miss fine grained detail. Defaults to 30.}

\item{min_dist}{Numeric between 0 and 0.99. min_dist controls how tightly points can be packed
together in 2D space. Lower values will generate more clumpy projections, but more accurately
preserve local structure.}

\item{metric}{distance metric for UMAP, defaults to pearson.}

\item{seed}{seed to generate reproducible UMAP projection. Defaults to no seed.}

\item{...}{}
}
\value{
fce object with PCA values added to reducedDims
}
\description{
See (https://umap-learn.readthedocs.io/en/latest/index.html) for
a detailed description of parameters.
}
