% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{calc_var_features}
\alias{calc_var_features}
\title{Get variable features}
\usage{
calc_var_features(fsce, expt = "rnaseq", n = 1000)
}
\arguments{
\item{fsce}{\code{\link{FunctionalSingleCellExperiment}}}

\item{expt}{Data to use for calculating variable features
(default is \code{rnaseq}). Must be present in \code{names(fsce)}.}

\item{n}{number of variable features to return}
}
\value{
Names of variable features
}
\description{
Identify features with the most variance in the data set.
}
\examples{
calc_var_features(fsce_small)[1:5]

calc_var_features(fsce_small, expt = "haircut")[1:5]
}
