% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{clr_normalize}
\alias{clr_normalize}
\title{Simple normalization for functional data using centered log ratio}
\usage{
clr_normalize(mat)
}
\arguments{
\item{mat}{input matrix}
}
\value{
matrix of normalized values. Normalization performed by dividing each
function value by the geometric mean of all functional values for a cell,
and returning log values
}
\description{
Simple normalization for functional data using centered log ratio
}
