% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_kmeans}
\alias{cluster_kmeans}
\title{Run k-means clustering algorithm}
\usage{
cluster_kmeans(fsce, expt = "rnaseq", k, method = "PCA",
  n_dims = NULL, seed = NULL, ...)
}
\arguments{
\item{fsce}{An object of class \code{\link{FunctionalSingleCellExperiment}}}

\item{expt}{Data to use for calculating variable features
(default is \code{rnaseq}). Must be present in \code{names(fsce)}.}

\item{k}{number of classes}

\item{method}{dimensionality reduction method for clustering (defaults to
PCA)}

\item{n_dims}{specify the number of dimensions from "dr" to use for
clustering, defaults to all dimensions}

\item{seed}{seed for reproducible result}

\item{...}{additional arguments to pass to \code{\link[stats:kmeans]{stats::kmeans()}}}
}
\value{
fsce with \code{k_cluster} in \code{expt} colData.
}
\description{
Run k-means clustering algorithm
}
\examples{
# calculate PCA for k-means default method
fsce <- calc_pca(fsce_small)

fsce <- cluster_kmeans(fsce, k = 6)

SingleCellExperiment::colData(fsce[["rnaseq"]], "k_cluster")

}
\seealso{
Other clustering functions: \code{\link{cluster_leiden}}
}
\concept{clustering functions}
