% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_leiden}
\alias{cluster_leiden}
\title{Run Leiden clustering algorithm}
\source{
\url{https://github.com/vtraag/leidenalg}

\url{https://github.com/TomKellyGenetics/leiden}
}
\usage{
cluster_leiden(fsce, expt = "rnaseq", method = "PCA", dims = 1:5,
  prune = 1/15, seed = NULL,
  partition_type = "ModularityVertexPartition", ...)
}
\arguments{
\item{fsce}{An object of class \code{\link{FunctionalSingleCellExperiment}}}

\item{expt}{Data to use for calculating variable features
(default is \code{rnaseq}). Must be present in \code{names(fsce)}.}

\item{method}{dimensionality reduction method for clustering (defaults to
PCA)}

\item{dims}{dimensions to use for nearest-neighbor calculation}

\item{prune}{Pruning parameter for shared nearest-neighbor calculation.}

\item{seed}{seed for \code{leidenalg$find_partition()}}

\item{partition_type}{partitioning algorithm (see \code{\link[leiden:leiden]{leiden::leiden}}).
(defaults to "ModularityVertexPartition")}

\item{...}{Parameters to pass to the Python \code{leidenalg} function.}
}
\value{
fsce with \code{leiden_cluster} in \code{expt} colData.
}
\description{
Runs Leiden community detection algorithm to detect clusters.
}
\details{
Execute \code{\link[=install_py_deps]{install_py_deps()}} to install required python modules \code{leidenalg}
and \code{igraph}.
}
\examples{
\donttest{
fsce_small <- cluster_leiden(fsce_small)
SingleCellExperiment::colData(fsce_small[["rnaseq"]])
}

}
\seealso{
Other clustering functions: \code{\link{cluster_kmeans}}
}
\concept{clustering functions}
