% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsce.R
\name{create_fsce}
\alias{create_fsce}
\title{Create a functional single-cell experiment}
\usage{
create_fsce(expt_list)
}
\arguments{
\item{expt_list}{a list mapping experiment names to \code{SingleCellExperiments}
created by an \code{sce_} function.}
}
\value{
\code{MultiAssayExperiment} containing one or more \code{SingleCellExperiment}.
}
\description{
A functional single-cell experiment (\code{fsce}) is implemented as a
\code{MultiAssayExperiment} with one or more \code{SingleCellExperiment}s (\code{sce}). One
of these is a single-cell mRNA sequencing experiment that defines mRNA counts
for single cells. Other \code{sce}s include functional data from haircut.
}
\examples{

\dontrun{
fsce <- create_fsce(
  list(
    rnaseq  = sce_rnaseq(scrunchy_data("mrna.csv.gz")),
    haircut = sce_haircut(scrunchy_data("haircut.csv.gz"))
  )
)

fsce
}

}
