% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_haircut}
\alias{create_haircut}
\title{Create a haircut object as a MultiAssayExperiment}
\usage{
create_haircut(rna_data, haircut_data, id_from_name = TRUE,
  id_delim = ".", id_fields = 2)
}
\arguments{
\item{rna_data}{UMI count matrix}

\item{haircut_data}{Functional count matrix}

\item{id_from_name}{Extract sample id from cell name (default = TRUE)
If false then the sample_id field will be populated with NA.}

\item{id_delim}{delimiter to split cell name (default = ".")}

\item{id_fields}{index(es) of fields to extract from name for determining sample id from name
(defaults to c(2))}
}
\description{
Create a haircut object as a MultiAssayExperiment
}
