% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{create_sce_feature}
\alias{create_sce_feature}
\title{Create a single-cell feature experiment}
\usage{
create_sce_feature(path, norm_method = "clr_normalize")
}
\arguments{
\item{path}{path to output matrices, or R matrix object.}

\item{norm_method}{Normalization method for \code{counts}. Normalized data
is stored in \code{logcounts}. Set to \code{NULL} to skip normalization.}
}
\value{
\code{SingleCellExperiment} containing a \code{sparseMatrix} of counts
}
\description{
Create a single-cell feature experiment
}
