% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{create_sce_haircut}
\alias{create_sce_haircut}
\title{Create a single-cell Haircut experiment}
\usage{
create_sce_haircut(path, norm_method = "clr_normalize", adducts = NULL)
}
\arguments{
\item{path}{path to output matrices.}

\item{norm_method}{Normalization method for \code{counts}. Normalized data
is stored in \code{logcounts}. Set to \code{NULL} to skip normalization.}

\item{adducts}{\code{data_frame} with positions of hairpin adducts. Expects
two columns named \code{adduct} and \code{pos}.}
}
\value{
\code{SingleCellExperiment} containing a \code{sparseMatrix} of counts
}
\description{
Create a single-cell Haircut experiment
}
\examples{
create_sce_haircut(scrunchy_data("haircut/"))

}
