% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{create_sce_rnaseq}
\alias{create_sce_rnaseq}
\title{Create a single-cell mRNA-seq experiment}
\usage{
create_sce_rnaseq(path, norm_method = "log_normalize")
}
\arguments{
\item{path}{path to output matrices.}

\item{norm_method}{Normalization method for \code{counts}. Normalized data
is stored in \code{logcounts}. Set to \code{NULL} to skip normalization.}
}
\value{
\code{SCERnaSeq} containing a \code{sparseMatrix} of counts
}
\description{
Create a single-cell mRNA-seq experiment
}
\examples{
create_sce_rnaseq(scrunchy_data("mrna/"))
}
