% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsce.R
\name{expt_haircut}
\alias{expt_haircut}
\title{Create a single-cell Haircut experiment}
\usage{
expt_haircut(csv, adducts)
}
\arguments{
\item{csv}{path to CSV counts matrix}

\item{adducts}{\code{data_frame} with positions of hairpin adducts. Expects
two columns named \code{adduct} and \code{pos}.}
}
\value{
\code{SingleCellExperiment} containing a \code{matrix} of counts
}
\description{
Create a single-cell Haircut experiment
}
\examples{
\dontrun {
expt_haircut(scrunchy_data("haircut.csv.gz"))
}

}
