% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsce.R
\name{expt_rnaseq}
\alias{expt_rnaseq}
\title{Create single-cell mRNA-seq experiment}
\usage{
expt_rnaseq(csv)
}
\arguments{
\item{csv}{path to CSV counts matrix}
}
\value{
\code{SingleCellExperiment} containing a \code{sparseMatrix} of counts
}
\description{
Create single-cell mRNA-seq experiment
}
\examples{
\dontrun {
expt_rnaseq(scrunchy_data("mrna.csv.gz"))
}

}
