% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{filter_cells}
\alias{filter_cells}
\title{Filter a data subset for specific cells}
\usage{
filter_cells(x, cell_ids)
}
\arguments{
\item{x}{a data subset}

\item{cell_ids}{vector of cell ids}
}
\description{
Expects a column named \code{cell_id} in the the data subset.
}
\examples{
x <- subset_dim(fsce_small, expt = "rnaseq", dim_name = "UMAP")

filter_cells(x, c("TGCGGGTTCAACCATG.1", "TGGTTCCTCACTTACT.1"))

}
