% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-barcodes.R
\name{filter_molecules}
\alias{filter_molecules}
\title{Filter barcodes in molecule file and write to disk}
\usage{
filter_molecules(molecule_file, output_file, bcs_to_keep)
}
\arguments{
\item{molecule_file}{path molecule.tsv.gz flatfile produced by scrunchy
pipeline}

\item{output_file}{output file name, will be compressed by default.}

\item{bcs_to_keep}{character vector of barcode sequences to keep in
output}
}
\description{
Useful for only keeping cell associated molecules for example.
}
