% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fsce_small}
\alias{fsce_small}
\title{An example FunctionalSingleCellExperiment}
\format{An object of class \code{FunctionalSingleCellExperiment} of length 2.}
\usage{
fsce_small
}
\description{
There are two \code{SingleCellExperiments}:
}
\details{
\enumerate{
\item \code{rnaseq}: a \code{SCERnaSeq} object containing a single-cell mRNA sequencing
(10x Genomics V2 3´ ) cell/gene counts matrix in the \code{counts} slot.
Log-normalized counts are in the \code{logcounts} slot.
\item \code{haircut}: a \code{SCEHaircut} object containing a Haircut experiment using DNA
repair oligos. Counts are in the \code{counts} slot, and centered log-ratio
normalized counts are in the \code{logcounts} slot.
}

Repair substrates include:
\itemize{
\item A "normal" DNA oligo (\code{Normal_})
\itemize{
\item a ribonucleotide (rG) at position 45 (\code{Ribo_})
\item an Abasic site at position 45 (\code{Abasic_})
\item a uracil:adenosine base-pair, with uracil at position 45 (\code{Uracil_})
\item a uracil:guanosine base-pair, with uracil at position 45 (\code{GU_})
}
}

Generated by \code{data-raw/fsce_small.R}.
}
\keyword{datasets}
