% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fsce_tidy}
\alias{fsce_tidy}
\title{Tidied tibble of data from fsce_small}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 250 rows and 6 columns.}
\source{
Generated by \code{data-raw/fsce_tidy.R}.
}
\usage{
fsce_tidy
}
\description{
Used for function examples.
}
\details{
Variables include:
\itemize{
\item \code{cell_id},
\item UMAP dimensions (\code{UMAP1} and \code{UMAP2})
\item k-means clusters (\code{k_cluster})
\item Top ten most variable activities (\code{Uracil_45}, \code{riboG_44}, etc.)
}
}
\examples{
fsce_tidy
}
\keyword{datasets}
