% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsce.R
\docType{class}
\name{FunctionalSingleCellExperiment-class}
\alias{FunctionalSingleCellExperiment-class}
\title{An S4 class to represent a functional single cell experiment}
\description{
An S4 class to represent a functional single cell experiment
}
