% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsce.R
\name{FunctionalSingleCellExperiment}
\alias{FunctionalSingleCellExperiment}
\title{Constructor for a \code{FunctionalSingleCellExperiment} object}
\usage{
FunctionalSingleCellExperiment(...)
}
\arguments{
\item{...}{params for \code{\link[MultiAssayExperiment:MultiAssayExperiment]{MultiAssayExperiment::MultiAssayExperiment}}}
}
\description{
A thin wrapper around \code{MultiAssayExperiment::MultiAssayExperiment}.
}
