% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_features.R
\name{get_var_features}
\alias{get_var_features}
\title{Get variable features}
\usage{
get_var_features(fsce, expt = "rnaseq", n = 1000)
}
\arguments{
\item{fsce}{fsce object}

\item{expt}{Data to use for calculating variable features
(default is \code{rnaseq}). Must be present in \code{names(fsce)}.}

\item{n}{number of variable features to return}
}
\value{
Names of variable features
}
\description{
Identify features with the most variance in the data set.
}
