% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vargenes.R
\name{get_var_genes}
\alias{get_var_genes}
\title{Get variable genes}
\usage{
get_var_genes(fce, expt = "sce", n_genes = 1000)
}
\arguments{
\item{fce}{fce object}

\item{expt}{Data to use for calculating variable genes,
one of either sce (rna data, the default) or fsce (functional data)}

\item{n_genes}{n variable genes to return}
}
\value{
Character vecctor
}
\description{
Get variable genes
}
