% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haircut.R
\name{hairpin_coverage}
\alias{hairpin_coverage}
\title{Calculate coverage across hairpins}
\usage{
hairpin_coverage(fsce, expt = "haircut", cell_ids = NULL)
}
\arguments{
\item{fsce}{\code{\link{FunctionalSingleCellExperiment}}}

\item{expt}{Data to use for calculating variable features
(default is \code{rnaseq}). Must be present in \code{names(fsce)}.}

\item{cell_ids}{cell ids to include in coverage calculation}
}
\description{
Calculate coverage across hairpins
}
\examples{
hairpin_coverage(fsce_small)

}
