% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_gene_ids}
\alias{human_gene_ids}
\title{Human gene identifiers}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 33514 rows and 2 columns.}
\source{
Generated by \code{data-raw/human-gene-ids.R}
}
\usage{
human_gene_ids
}
\description{
Used for \code{\link[=calc_cell_cycle]{calc_cell_cycle()}}
}
\examples{
human_gene_ids

}
\keyword{datasets}
