% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{install_py_deps}
\alias{install_py_deps}
\title{Install python dependencies for \code{cluster_leiden}.}
\usage{
install_py_deps(method = "auto", conda = "auto")
}
\arguments{
\item{method}{method param for \code{\link[reticulate:py_install]{reticulate::py_install()}}}

\item{conda}{conda param for \code{\link[reticulate:py_install]{reticulate::py_install()}}}
}
\description{
Users must run this prior to using \code{cluster_leiden()}.
}
