% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodes-utils.R
\name{is.molecule_df}
\alias{is.molecule_df}
\title{Test if the object is a molecule_df.}
\usage{
is.molecule_df(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{\link[=molecule_df]{molecule_df()}} class.
}
\description{
Test if the object is a molecule_df.
}
