% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_matrix_csv}
\alias{load_matrix_csv}
\title{Load count matrix in CSV format}
\usage{
load_matrix_csv(csv)
}
\arguments{
\item{csv}{path to CSV file}
}
\value{
\code{data.frame} with cells as row.names
}
\description{
Generated by scrunchy pipeline
}
