% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodes-utils.R
\name{molecule_df}
\alias{molecule_df}
\title{data.frame that contains molecule info}
\usage{
molecule_df(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{\link[=molecule_df]{molecule_df()}} class.
}
\description{
data.frame that contains molecule info
}
