% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\docType{methods}
\name{normalize}
\alias{normalize}
\alias{normalize,SingleCellExperiment-method}
\title{Normalize data in a SingleCellExperiment}
\usage{
normalize(x, method, ...)

\S4method{normalize}{SingleCellExperiment}(x, method = "log_normalize",
  ...)
}
\arguments{
\item{x}{matrix to normalize}

\item{method}{normalization method}

\item{...}{addition arguments for normalization method}
}
\description{
Normalize data in a SingleCellExperiment
}
