% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_fce}
\alias{normalize_fce}
\title{Normalize single cell rna and functional data}
\usage{
normalize_fce(fce, rna_method = "log_normalize",
  functional_method = "clr")
}
\arguments{
\item{fce}{fce object generted by \code{\link[=create_fce]{create_fce()}}}

\item{rna_method}{normalization method for RNA data}

\item{functional_method}{normalization method for functional data}
}
\value{
fce object with log normalized counts returned in the logcounts slot
}
\description{
Normalize single cell rna and functional data
}
