% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_marker_genes}
\alias{pbmc_marker_genes}
\title{Comman human PBMC marker genes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 3 columns.}
\usage{
pbmc_marker_genes
}
\description{
Comman human PBMC marker genes
}
\examples{
pbmc_marker_genes

}
\keyword{datasets}
