% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_activity}
\alias{plot_activity}
\title{Plot activities per cluster}
\usage{
plot_activity(data, activity, cluster)
}
\arguments{
\item{data}{data to plot}

\item{activity}{activity variable}

\item{cluster}{cluster variable}
}
\description{
Plot activities per cluster
}
