% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cells}
\alias{plot_cells}
\title{Plot cells in a reduced dimensionality space}
\usage{
plot_cells(fce, features, ...)
}
\arguments{
\item{fce}{fce object of class MultiAssayExperiment}

\item{features}{features to plot}

\item{...}{additional arguments for \code{plot_feature()}}
}
\description{
Plot cells in two dimensions. Cells can be colored using \code{features}, which
specifies gene names or columns in \code{colData}.
}
