% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_dims}
\alias{plot_dims}
\title{Scatter plot of cells in a two-dimensional embedding.}
\usage{
plot_dims(df, x, y, color = NULL, size = 0.1, alpha = 1,
  palette = "cloupe", title = NULL)
}
\arguments{
\item{df}{plot data}

\item{x}{variable for x-axis}

\item{y}{variable for y-axis}

\item{color}{variable for point colors (default is black)}

\item{size}{size for \code{\link{geom_point}}}

\item{alpha}{alpha for \code{\link{geom_point}}}

\item{palette}{palette for continuous colors. One of cloupe (the default),
brewer, viridis.}

\item{title}{plot title}
}
\description{
This is the base plot for superimposing annotations like cell types,
cluster assignments, and measured activities.
}
\details{
Embeddings can be calculated by \code{\link[=calc_umap]{calc_umap()}} and \code{\link[=calc_tsne]{calc_tsne()}}, and
retrieved with \code{\link[=tidy_dims]{tidy_dims()}}.
}
\examples{
plot_dims(fsce_tidy, UMAP1, UMAP2, size = 1)

plot_dims(fsce_tidy, UMAP1, UMAP2, Uracil_45, size = 1)

plot_dims(fsce_tidy, UMAP1, UMAP2, k_cluster, size = 1)


}
