% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_feature}
\alias{plot_feature}
\title{Plot cells in reduced dimensionality 2D space}
\usage{
plot_feature(fce, feature = NULL, expt = "sce", method = "UMAP",
  plot_dat = NULL, pt_size = 0.01, pt_alpha = 1,
  label_text = FALSE, label_size = 6, label_color = "grey",
  .cols = NULL, cell_filter = NULL, palette_type = "cloupe",
  col_pal = "Reds", max_y = NULL, plot_title = NULL,
  dims_to_plot = c(1, 2))
}
\arguments{
\item{fce}{fce object of class MultiAssayExperiment}

\item{feature}{single feature to plot, either gene name or column in colData}

\item{expt}{Data type to use for plot, one of either sce (rna data, the default)
or fsce (functional data). Defaults to "sce"}

\item{method}{dimensionality reduction for plotting. defaults to UMAP}

\item{plot_dat}{supplemental data.frame containing feature to plot.
Must have a column named cell that contains matching colnames in colData}

\item{pt_size}{size of points produced by geom_point}

\item{pt_alpha}{alpha value for points plotted by geom_point}

\item{label_text}{if TRUE display feature labels on plot}

\item{label_size}{size of label text}

\item{label_color}{color of label text}

\item{.cols}{vector of colors to use for plot.}

\item{cell_filter}{character vector of cell names to include in plot}

\item{palette_type}{color palette type to use (either viridis, brewer, or cloupe)
defaults to using cellranger loupe-like colors}

\item{col_pal}{palette name to use if palette_type is brewer}

\item{max_y}{maximum feature value to set scale to. Defaults to max of the feature}

\item{plot_title}{string to supply for title for the plot, otherwise title is the feature}

\item{dims_to_plot}{number of dimensions}
}
\description{
Cells can be colored by gene or feature in meta.data dataframe
}
