% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_hairpin}
\alias{plot_hairpin}
\title{Calculate and plot coverage across hairpins}
\usage{
plot_hairpin(hcut_obj, cells_to_plot = NULL, return_df = FALSE, ...)
}
\arguments{
\item{hcut_obj}{haircut_object generated by create_haircut}

\item{cells_to_plot}{vector of cells ids to include in coverage calculation}

\item{return_df}{return coverage values as data.frame (default = FALSE)}

\item{...}{additional arguments to pass to \code{\link{plot_hairpin_coverage}}}
}
\description{
Calculate and plot coverage across hairpins
}
