% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_10x_matrix}
\alias{read_10x_matrix}
\title{Load 10x mtx matrix as sparseMatrix}
\usage{
read_10x_matrix(path, cell_prefix = NULL, strip_10x_suffix = TRUE,
  use_gene_symbols = TRUE)
}
\arguments{
\item{path}{path to 10x directory with mtx matrix}

\item{cell_prefix}{string to prefix to cell_id  (default = NULL)}

\item{strip_10x_suffix}{remove numeric suffix added by 10x
e.g. change "TTTGTCAAGGTGTGGT-1" to "TTTGTCAAGGTGTGGT" (default = TRUE)}

\item{use_gene_symbols}{If TRUE use gene symbols as row.names, if false then
gene_ids will be used (default = TRUE)}
}
\description{
Load 10x mtx matrix as sparseMatrix
}
