% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_matrix}
\alias{read_matrix}
\title{Load mtx formatted 10x or haircut matrices}
\usage{
read_matrix(path, cell_prefix = NULL, strip_10x_suffix = TRUE,
  use_gene_symbols = TRUE, split_matrix = FALSE,
  matrix_fn = "matrix.mtx.gz", features_fn = "features.tsv.gz",
  barcodes_fn = "barcodes.tsv.gz")
}
\arguments{
\item{path}{path to directory with mtx matrix}

\item{cell_prefix}{string to prefix to cell_id  (default = NULL)}

\item{strip_10x_suffix}{remove numeric suffix added by 10x e.g. change
"TTTGTCAAGGTGTGGT-1" to "TTTGTCAAGGTGTGGT" (default = TRUE)}

\item{use_gene_symbols}{If TRUE use gene symbols as row.names, if false then
gene_ids will be used (default = TRUE)}

\item{split_matrix}{Return a list of matrices split by feature types. Requires
cellranger v.3 output.}

\item{matrix_fn}{filename for matrix}

\item{features_fn}{filename for genes file}

\item{barcodes_fn}{filename for barcodes file}
}
\description{
Load mtx formatted 10x or haircut matrices
}
