% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodes-utils.R
\name{read_molecules}
\alias{read_molecules}
\title{Read molecule file into memory}
\usage{
read_molecules(molecule_file, return_ids = TRUE)
}
\arguments{
\item{molecule_file}{molecule.tsv.gz flatfile produced by scrunchy pipeline}

\item{return_ids}{convert barcodes and features to names rather than indexes in
barcodes.tsv and features.tsv (default = TRUE)}
}
\description{
Read molecule file into memory
}
