% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{run_kmeans}
\alias{run_kmeans}
\title{Perform Kmeans clustering}
\usage{
run_kmeans(fce, k, expt = "sce", dr = "PCA", n_dims = NULL, ...)
}
\arguments{
\item{fce}{fce object}

\item{k}{specify k number of classes}

\item{expt}{experiment to cluster (either "sce" or "fsce, defaults to "sce")}

\item{dr}{which dimensionality reduction to use for clustering (defaults to PCA)}

\item{n_dims}{specify the number of dimensions from "dr" to use for clustering, defaults to all dimensions}

\item{...}{additional arguments to pass to \code{\link[stats:kmeans]{stats::kmeans()}}}
}
\description{
Perform Kmeans clustering
}
