% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{sce_haircut}
\alias{sce_haircut}
\title{Create a single-cell Haircut experiment}
\usage{
sce_haircut(csv, norm_method = "clr_normalize", adducts = NULL)
}
\arguments{
\item{csv}{path to CSV counts matrix}

\item{norm_method}{Normalization method for \code{counts}. Normalized data
is stored in \code{logcounts}. Set to \code{NULL} to skip normalization.}

\item{adducts}{\code{data_frame} with positions of hairpin adducts. Expects
two columns named \code{adduct} and \code{pos}.}
}
\value{
\code{SCEHaircut} containing a \code{matrix} of counts
}
\description{
Create a single-cell Haircut experiment
}
\examples{
\dontrun {
sce_haircut(scrunchy_data("haircut.csv.gz"))
}

}
