% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{sce_rnaseq}
\alias{sce_rnaseq}
\title{Create a single-cell mRNA-seq experiment}
\usage{
sce_rnaseq(csv, norm_method = "log_normalize")
}
\arguments{
\item{csv}{path to CSV counts matrix}

\item{norm_method}{Normalization method for \code{counts}. Normalized data
is stored in \code{logcounts}. Set to \code{NULL} to skip normalization.}
}
\value{
\code{SCERnaSeq} containing a \code{sparseMatrix} of counts
}
\description{
Create a single-cell mRNA-seq experiment
}
\examples{
\dontrun {
sce_rnaseq(scrunchy_data("mrna.csv.gz"))
}

}
