% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrunchy-package.R
\docType{package}
\name{scrunchy}
\alias{scrunchy}
\alias{scrunchy-package}
\title{scrunchy: single-cell analysis of functional heterogeneity}
\description{
scrunchy provides tools to analyze and visualize data from
single-cell assays for functional heterogeneity.
}
\details{
To learn more about scrunchy, start with the vignette:
\code{browseVignettes(package = "scrunchy")}
}
\seealso{
Report bugs at \url{https://github.com/hesselberthlab/scrunchy/issues}
}
\author{
Jay Hesselberth \href{mailto:jay.hesselberth@gmail.com}{jay.hesselberth@gmail.com}

Kent Riemondy \href{mailto:kent.riemondy@gmail.com}{kent.riemondy@gmail.com}

Amanda Richer \href{mailto:mandy.l.richer@gmail.com}{mandy.l.richer@gmail.com}
}
