% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stat_anova_grouped}
\alias{stat_anova_grouped}
\title{Analysis of variance of activities across groups}
\usage{
stat_anova_grouped(tbl, group = NULL)
}
\arguments{
\item{tbl}{tidied data from a \code{SingleCellExperiment}}

\item{group}{variable for generating combinations}
}
\description{
Analysis of variance of activities across groups
}
\examples{
x <- fsce_tidy[c("k_cluster", "Uracil_45", "riboG_44")]
x$k_cluster <- as.factor(x$k_cluster)

## default is list of aov models
stat_anova_grouped(x, k_cluster)

}
