% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_col}
\alias{subset_col}
\title{Retrieve colData from an fsce}
\usage{
subset_col(fsce, expt = "rnaseq", col_names = NULL)
}
\arguments{
\item{fsce}{\code{\link{FunctionalSingleCellExperiment}}}

\item{expt}{Data to use for calculating variable features
(default is \code{rnaseq}). Must be present in \code{names(fsce)}.}

\item{col_names}{column names to retrieve}
}
\value{
Data in \code{tibble} format.
}
\description{
Subset is returned in a tidied, tibble format, suitable for plotting or analysis.
}
\details{
Cell IDs are added in column \code{cell_id} to facilitate joining of subsets.
}
\examples{
subset_col(fsce_small, expt = "haircut", col_names = c("Uracil_45"))

}
