% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_counts}
\alias{subset_counts}
\title{Retrieve counts from an fsce}
\usage{
subset_counts(fsce, expt, vars = NULL)
}
\arguments{
\item{fsce}{\code{\link{FunctionalSingleCellExperiment}}}

\item{expt}{Data to use for calculating variable features
(default is \code{rnaseq}). Must be present in \code{names(fsce)}.}

\item{vars}{variables to retrieve (default: all)}
}
\description{
Subset is returned in a tidied, tibble format, suitable for plotting or analysis.
}
\details{
Cell IDs are added in column \code{cell_id} to facilitate joining of subsets.
}
\examples{
subset_counts(fsce_small, expt = "haircut", vars = c("Uracil_45"))

}
