% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_dim}
\alias{subset_dim}
\title{Retrieve reducedDim data}
\usage{
subset_dim(fsce, expt = "rnaseq", dim_name = "UMAP", dims = c(1, 2))
}
\arguments{
\item{fsce}{\code{\link{FunctionalSingleCellExperiment}}}

\item{expt}{Data to use for calculating variable features
(default is \code{rnaseq}). Must be present in \code{names(fsce)}.}

\item{dim_name}{name of reducedDim to retrieve (default: UMAP)}

\item{dims}{dimensions to retrieve (default: \code{c(1, 2)})}
}
\value{
Data in \code{tibble} format.
}
\description{
Subset is returned in a tidied, tibble format, suitable for plotting or
analysis.
}
\details{
Cell IDs are added in column \code{cell_id} to facilitate joining of subsets.
}
\examples{
subset_dim(fsce_small, expt = "rnaseq", dim_name = "UMAP")

subset_dim(fsce_small, expt = "rnaseq", dim_name = "TSNE")

subset_dim(fsce_small, expt = "rnaseq", dim_name = "PCA", dims = c(1:3))

}
