% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy_all}
\alias{tidy_all}
\title{Tidy all}
\usage{
tidy_all(fsce, dimnames = NULL, dims = c(1, 2), genes = NULL,
  repair = NULL)
}
\arguments{
\item{fsce}{An object of class \code{\link{FunctionalSingleCellExperiment}}.}

\item{dimnames}{vector of dimnames to retrieve. If \code{NULL}, retrieve all
dimnames.}

\item{dims}{vector of dimensions to retrieve}

\item{genes}{vector of genes to retrieve. If \code{NULL}, retrieve all genes.}

\item{repair}{vector of repair sites to retrieve. If \code{NULL} retrieve all reapir sites.}
}
\description{
Tidy all
}
\section{Tidying}{

This is a data tidier for a \code{\link{FunctionalSingleCellExperiment}}. Returns data
from \code{\link{SingleCellExperiment}} in a tidy format, where variables are columns
and observations are rows.

If the \code{\link{FunctionalSingleCellExperiment}} contains more than one
\code{\link{SingleCellExperiment}}, data from each \code{sce} are joined using the \code{cell_id}
variable name, and a new \code{experiment} column contains the name of the source
\code{sce}.
}

\examples{
tidy_all(fsce_small, dimnames = c("UMAP"), genes = c("TP53"), repair = c("Uracil_45"))

}
\seealso{
Other tidiers: \code{\link{tidy_coldata}},
  \code{\link{tidy_counts}}, \code{\link{tidy_dims}},
  \code{\link{tidy_logcounts}}
}
\concept{tidiers}
