% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy_counts}
\alias{tidy_counts}
\title{Tidy counts data}
\usage{
tidy_counts(fsce)
}
\arguments{
\item{fsce}{A \code{\link{FunctionalSingleCellExperiment}}.}
}
\description{
Tidy counts data
}
\section{Tidying}{

This method is a data tidier for a \code{\link{FunctionalSingleCellExperient}}. Returns
data from an \code{fsce} in a tidy format, where variables are columns and
observations are rows. Output will contain \code{experiment} and \code{cell_id} columns
that describe the source of the data.
}

\examples{
x <- fsce_small[ c("Uracil_45"), , "haircut"]
tidy_counts(x)

}
