% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy_counts_full}
\alias{tidy_counts_full}
\title{Tidy counts and return long format}
\usage{
tidy_counts_full(fsce, expt = "haircut")
}
\arguments{
\item{fsce}{An object of class \code{\link{FunctionalSingleCellExperiment}}.}
}
\description{
Tidy counts and return long format
}
\section{Tidying}{

This is a data tidier for a \code{\link{FunctionalSingleCellExperiment}}. Returns data
from \code{\link{SingleCellExperiment}} in a tidy format, where variables are columns
and observations are rows.

If the \code{\link{FunctionalSingleCellExperiment}} contains more than one
\code{\link{SingleCellExperiment}}, data from each \code{sce} are joined using the \code{cell_id}
variable name, and a new \code{experiment} column contains the name of the source
\code{sce}.
}

\examples{
tidy_counts_full(fsce_small[1:10, , "haircut"])
tidy_counts_full(fsce_small[1:10, , "rnaseq"])

}
\seealso{
Other tidiers: \code{\link{tidy_coldata}},
  \code{\link{tidy_counts}}, \code{\link{tidy_dims}},
  \code{\link{tidy_logcounts_full}},
  \code{\link{tidy_logcounts}}, \code{\link{tidy_rowdata}}
}
\concept{tidiers}
