% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy_dims}
\alias{tidy_dims}
\title{Tidy reducedDims data}
\usage{
tidy_dims(sce, dimnames = NULL, dims = c(1, 2))
}
\arguments{
\item{sce}{An object of class \code{\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment::SingleCellExperiment}}.}

\item{dimnames}{vector of dimnames to retrieve. If \code{NULL}, retrieve all
dimnames.}

\item{dims}{vector of dimensions to retrieve}
}
\description{
Tidy reducedDims data
}
\section{Tidying}{

This is a data tidier for a \code{SingleCellExperient}. Returns
data from an \code{sce} in a tidy format, where variables are columns and
observations are rows.
}

\examples{
tidy_dims(fsce_small)

tidy_dims(fsce_small, dimnames = c("UMAP"))
}
\seealso{
Other tidiers: \code{\link{tidy_coldata}},
  \code{\link{tidy_counts}}, \code{\link{tidy_logcounts}}
}
\concept{tidiers}
