% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{tidy_stats_grouped}
\alias{tidy_stats_grouped}
\title{Tidy grouped statistics}
\usage{
tidy_stats_grouped(x, id = "activity")
}
\arguments{
\item{x}{list of named statistical results, e.g. from \code{\link{stat_anova_grouped}}.}

\item{id}{name of id variable in the output (default: "activity")}
}
\description{
Tidy grouped statistics
}
\examples{
x <- fsce_tidy[c("k_cluster", "Uracil_45")]
x$k_cluster <- as.factor(x$k_cluster)

res <- stat_anova_grouped(x, k_cluster)
tidy_stats_grouped(res)

}
