% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{umitools_to_mtx}
\alias{umitools_to_mtx}
\title{Convert umitools flat format tsv to sparseMatrix .mtx format}
\usage{
umitools_to_mtx(count_file, output_path = NULL, ...)
}
\arguments{
\item{count_file}{path to umitools output file}

\item{output_path}{path for output files. matrix.mtx, barcodes.tsv and features.tsv
will be generated at the supplied path, or by default created in same directory as
the \code{count_file}.}

\item{...}{additional arguments to pass to \code{\link[readr:read_tsv]{readr::read_tsv()}}}
}
\description{
Convert umitools flat format tsv to sparseMatrix .mtx format
}
