% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_matrix}
\alias{write_matrix}
\title{Write a sparseMatrix to disk}
\usage{
write_matrix(mat, output_path)
}
\arguments{
\item{mat}{sparseMatrix}

\item{output_path}{path for output files. matrix.mtx.gz,
barcodes.tsv.gz and features.tsv.gz will be generated
at the supplied path}
}
\description{
Write a sparseMatrix to disk
}
