% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.R
\name{calc_pca}
\alias{calc_pca}
\title{Calculate PCs use irlba}
\usage{
calc_pca(fce, expt = "sce", assay = "logcounts", genes = NULL,
  n_pcs = 20, scale = TRUE)
}
\arguments{
\item{fce}{fce object}

\item{expt}{one of either sce (rna data, the default) or fsce (functional data)}

\item{assay}{select which assay data to use for PCA, defaults to logcounts
(i.e. log normalized data)}

\item{genes}{vector of genes to include in PCA, defaults to all genes}

\item{n_pcs}{number of principle components to return}

\item{scale}{perform PCA on input on scaled data, defaults to TRUE}
}
\value{
fce object with PCA values added to reducedDims
}
\description{
Calculate PCs use irlba
}
