% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.R
\name{calc_tsne}
\alias{calc_tsne}
\title{Generate 2D cell embeddings using tSNE}
\usage{
calc_tsne(fce, expt = "sce", dr = "PCA", n_dims = NULL,
  perplexity = 30, theta = 0.5, seed = NA, ...)
}
\arguments{
\item{fce}{fce object}

\item{expt}{Data to use for tSNE, one of either sce (rna data, the default) or fsce (functional data)}

\item{dr}{select which dimenality reduction data to use for tSNE, defaults to PCA}

\item{n_dims}{number of dimensions to pass to tSNE, defaults to all present in dr matrix}

\item{perplexity}{Defaults to 30.}

\item{theta}{Defaults to 0.5}

\item{seed}{seed to generate reproducible tSNE projection. Defaults to no seed.}

\item{...}{additional arguments to pass to \code{\link[Rtsne:Rtsne]{Rtsne::Rtsne()}}}
}
\value{
fce object with tSNE values added to reducedDims
}
\description{
See (https://github.com/jkrijthe/Rtsne) for
a detailed description of parameters.
}
