% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.R
\name{calc_tsne}
\alias{calc_tsne}
\title{Generate 2D cell embeddings using tSNE}
\usage{
calc_tsne(fsce, expt = "rnaseq", method = "PCA", n_dims = NULL,
  perplexity = 30, theta = 0.5, seed = NULL, ...)
}
\arguments{
\item{fsce}{\code{\link{FunctionalSingleCellExperiment}} object}

\item{expt}{Data to use for calculating variable features
(default is \code{rnaseq}). Must be present in \code{names(fsce)}.}

\item{method}{dimensionality reduction method for tSNE}

\item{n_dims}{number of dimensions for tSNE}

\item{perplexity}{tSNE perplexity value.}

\item{theta}{tSNE theta value.}

\item{seed}{seed for generate reproducible tSNE projection.}

\item{...}{additional arguments for \code{\link[Rtsne:Rtsne]{Rtsne::Rtsne()}}}
}
\value{
fce object with tSNE values added to reducedDims
}
\description{
Generate 2D cell embeddings using tSNE
}
\seealso{
See \code{\link[Rtsne:Rtsne]{Rtsne::Rtsne()}} for a detailed description of parameters.
}
