% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.R
\name{calc_umap}
\alias{calc_umap}
\title{Generate 2D cell embeddings using UMAP}
\usage{
calc_umap(fsce, expt = "rnaseq", method = "PCA", n_dims = NULL,
  n_neighbors = 30, min_dist = 0.3, metric = "euclidean",
  seed = NA, ...)
}
\arguments{
\item{fsce}{\code{\link{FunctionalSingleCellExperiment}} object}

\item{expt}{Data to use for calculating variable features
(default is \code{rnaseq}). Must be present in \code{names(fsce)}.}

\item{method}{dimenality reduction method to use for UMAP (default is "PCA")}

\item{n_dims}{number of dimensions to pass to UMAP, defaults to all present
in dr matrix}

\item{n_neighbors}{number of nearest neighbors to use for learning the
manifold. Low values will preserve local structure, at the expense missing
higher order organization. Higher values will capture more global structure
but miss fine grained detail. Defaults to 30.}

\item{min_dist}{Numeric between 0 and 0.99. min_dist controls how tightly
points can be packed together in 2D space. Lower values will generate more
clumpy projections, but more accurately preserve local structure.}

\item{metric}{distance metric for UMAP, defaults to pearson.}

\item{seed}{seed to generate reproducible UMAP projection. Defaults to no
seed.}

\item{...}{additional arguments for \code{\link[umap:umap]{umap::umap()}}}
}
\value{
\code{fsce} with UMAP values added to reducedDims
}
\description{
Generate 2D cell embeddings using UMAP
}
\seealso{
See \url{https://umap-learn.readthedocs.io} for a detailed description of
parameters.
}
