% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{clr_normalize}
\alias{clr_normalize}
\title{Centered log-ratio normalization}
\usage{
clr_normalize(mat)
}
\arguments{
\item{mat}{input matrix}
}
\value{
matrix of normalized values. Normalization performed by dividing each
value by the geometric mean of all values for a cell, returning log values.
}
\description{
This normalization strategy is used for CITE-seq and other feature data in
\href{https://satijalab.org/seurat/}{Seurat}.
}
\seealso{
https://stackoverflow.com/questions/2602583/geometric-mean-is-there-a-built-in
}
