% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fsce_small}
\alias{fsce_small}
\title{An example FunctionalSingleCellExperiment}
\format{An object of class \code{FunctionalSingleCellExperiment} of length 2.}
\source{
Generated by \code{data-raw/fsce_small.R}.
}
\usage{
fsce_small
}
\description{
Contains sample data for a functional single-cell experiment.
}
\details{
There are two \code{SingleCellExperiments}:
\enumerate{
\item \code{rnaseq}: a \code{SingleCellExperiment} object containing a single-cell mRNA sequencing
(10x Genomics V2 3 prime) cell/gene counts matrix in the \code{counts} slot.
Log-normalized counts are in the \code{logcounts} slot.
\item \code{haircut}: a \code{SingleCellExperiment} object containing a Haircut experiment using DNA
repair oligos. Counts are in the \code{counts} slot, and centered log-ratio
normalized counts are in the \code{logcounts} slot.
}

Haircut repair substrates include:
\itemize{
\item A "normal" DNA oligo (\code{Normal_})
\item a ribonucleotide (rG) at position 45 (\code{Ribo_})
\item an Abasic site at position 45 (\code{Abasic_})
\item a uracil:adenosine base-pair, with uracil at position 45 (\code{Uracil_})
\item a uracil:guanosine base-pair, with uracil at position 45 (\code{GU_})
}

Dimensionality reduction calculations are in the \code{rnaseq} experiment, including
\code{PCA}, \code{UMAP}, and \code{TSNE} slots. In addition, k-means clusters are available as
\code{rowData}.
}
\examples{
fsce_small

# Individual experiments
fsce_small[["rnaseq"]]

fsce_small[["haircut"]]

# Gene and activity names (first 5 items)
rownames(fsce_small[["rnaseq"]])[1:5]

rownames(fsce_small[["haircut"]])[1:5]

# subset with `[` using `[features, cell_ids, experiments]`
features <- c("Uracil_45", "TP53")
cell_ids <- c("GAAACTCTCCGAAGAG", "CGCTATCGTTCACCTC")

fsce_small[features, cell_ids, ]

library(SingleCellExperiment)
# dimensionality reduction results
reducedDimNames(fsce_small[["rnaseq"]])

# UMAP results (first 5 rows)
reducedDim(fsce_small[["rnaseq"]], "UMAP")[1:5, ]

# k-means cluster IDs
colData(fsce_small[["rnaseq"]])

}
\keyword{datasets}
