% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vargenes.R
\name{get_var_genes}
\alias{get_var_genes}
\title{Get variable genes}
\usage{
get_var_genes(fce, expt = "sce", n_genes = 1000)
}
\arguments{
\item{fce}{fce object}

\item{expt}{Data to use for calculating variable genes,
one of either \code{sce} (the default) or \code{fsce} (functional data)}

\item{n_genes}{number of variable genes to return}
}
\value{
Names of variable genes
}
\description{
Identify genes the most variance in the data set.
}
