% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{log_normalize}
\alias{log_normalize}
\title{Simple normalization for scrna-seq}
\usage{
log_normalize(mat, constant = 10000)
}
\arguments{
\item{mat}{input matrix}

\item{constant}{scalar to multiply normalized counts by to prevent
small numbers (1e5)}
}
\value{
matrix of normalized values. Normalization performed by dividing by
column sums (total counts per cell) and scaled by a scaling factor. Log
values are returned with a pseudocount of 1
}
\description{
Simple normalization for scrna-seq
}
