% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize single cell rna and functional data}
\usage{
normalize(hcut_obj, rna_method = "log_normalize", hcut_method = "clr")
}
\arguments{
\item{hcut_obj}{haircut object}

\item{rna_method}{normalization method for RNA data}

\item{hcut_method}{normalization method for functional data}
}
\value{
hcut_obj with log normalized counts returned in the logcounts slot
}
\description{
Normalize single cell rna and functional data
}
