% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_activity}
\alias{plot_activity}
\title{Plot activities per cluster}
\usage{
plot_activity(data, activity, group = NULL, labels = NULL,
  vertical = FALSE, stats = NULL, ...)
}
\arguments{
\item{data}{data to plot}

\item{activity}{activity variable}

\item{group}{grouping variable}

\item{labels}{legend labels}

\item{vertical}{should activity be on the y axis. Default is FALSE}

\item{stats}{dataframe to use to add p values on plot. Stats can be a dataframe output from \code{\link[=stat_activity_grouped]{stat_activity_grouped()}}.
Output must include columsn with names group, group1, q.value and/or p.value. If sepcificied, \code{vertical} is TRUE}

\item{...}{params for \code{\link[=add_stats]{add_stats()}}}
}
\description{
Generates a beeswarm plot of activity across specified groups
}
\examples{
plot_activity(fsce_tidy, Uracil_45, k_cluster)

plot_activity(fsce_tidy, riboG_44, k_cluster, labels = LETTERS[1:6])

plot_activity(fsce_tidy, Uracil_45, k_cluster, vertical = TRUE)

x <- fsce_tidy[c("k_cluster", "Uracil_45")]
stats <- stat_activity_grouped(x, group = k_cluster)
stats <- subset(stats, q.value < 0.01)
stats[,"y_loc"] = seq(max(x$Uracil_45), max(x$Uracil_45) + 3,  length.out = length(stats$group))
plot_activity(fsce_tidy, Uracil_45, k_cluster, stats = stats)

}
\seealso{
Other plot functions: \code{\link{plot_dims}},
  \code{\link{plot_hairpin_coverage}}
}
\concept{plot functions}
