% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_activity}
\alias{plot_activity}
\title{Plot activities per cluster}
\usage{
plot_activity(data, activity, group = NULL)
}
\arguments{
\item{data}{data to plot}

\item{activity}{activity variable}

\item{group}{grouping variable}
}
\description{
Generates a beeswarm plot of activity across specified groups
}
\examples{
cowplot::plot_grid(
  plotlist = list(
    plot_dims(fsce_tidy, UMAP1, UMAP2, k_cluster),
    plot_activity(fsce_tidy, Uracil_45, k_cluster),
    plot_activity(fsce_tidy, riboG_44, k_cluster)
  )
)

}
\seealso{
Other plotting: \code{\link{plot_dims}},
  \code{\link{plot_heatmap}}
}
\concept{plotting}
