% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_activity}
\alias{plot_activity}
\title{Plot activities per cluster}
\usage{
plot_activity(data, activity, group = NULL, labels = NULL)
}
\arguments{
\item{data}{data to plot}

\item{activity}{activity variable}

\item{group}{grouping variable}

\item{labels}{legend labels}
}
\description{
Generates a beeswarm plot of activity across specified groups
}
\examples{
plot_activity(fsce_tidy, Uracil_45, k_cluster),

plot_activity(fsce_tidy, riboG_44, k_cluster, labels = LETTERS[1:6])

}
\seealso{
Other plot functions: \code{\link{plot_dims}}
}
\concept{plot functions}
