% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt-haircut.R
\name{plot_hairpin_coverage}
\alias{plot_hairpin_coverage}
\title{Plot coverage across hairpins}
\usage{
plot_hairpin_coverage(fsce, expt = "haircut", cells = NULL,
  color = "hairpin")
}
\arguments{
\item{fsce}{\code{\link{FunctionalSingleCellExperiment}}}

\item{expt}{Data to use for calculating variable features
(default is \code{rnaseq}). Must be present in \code{names(fsce)}.}

\item{color}{variable to use for coloring lines (defaults to "hairpin")}
}
\description{
Plot coverage across hairpins
}
\examples{
plot_hairpin_coverage(fsce_small) + ggplot2::facet_wrap(~hairpin)
}
