% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_hairpin_coverage}
\alias{plot_hairpin_coverage}
\title{Plot coverage across hairpins}
\usage{
plot_hairpin_coverage(df, x, y, col = "sample", x_lab = "Position",
  y_lab = "Counts", pal = "Set1", point = F,
  facet_by_hairpin = TRUE, pt_alpha = 0.7, pt_size = 0.8,
  xlimits = NULL)
}
\arguments{
\item{df}{dataframe with coverage data}

\item{x}{column to use for x-axis}

\item{y}{column to use for y-axis}

\item{col}{column to use for coloring lines (defaults to "Sample")}

\item{x_lab}{label for x axis (defaults to "Position")}

\item{y_lab}{label for y axis (defaults to "Counts")}

\item{pal}{RcolorBrewer palette to use for coloring col (defaults to "Set1")}

\item{point}{show points in addition to lines (defaults to FALSE)}

\item{facet_by_hairpin}{plot each hairpin individually (defaults to TRUE)}

\item{pt_alpha}{passed to pt.alpha argument for \code{\link[ggplot2:geom_line]{ggplot2::geom_line()}} (defaults to 0.7)}

\item{pt_size}{passed to pt.size argument for \code{\link[ggplot2:geom_line]{ggplot2::geom_line()}} (defaults to 0.8)}

\item{xlimits}{two element numeric vector to supply to \code{\link[ggplot2:xlim]{ggplot2::xlim()}} (defaults to NULL)}
}
\description{
Plot coverage across hairpins
}
