% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stat_activity_grouped}
\alias{stat_activity_grouped}
\title{Compare activity differences between groups.}
\usage{
stat_activity_grouped(tbl, group, complete = FALSE, ...)
}
\arguments{
\item{tbl}{data from a \code{SingleCellExperiment}}

\item{group}{variable for generating combinations}

\item{complete}{If \code{TRUE}, generate complete group combinations (useful for
e.g., matrix visulatization of p-values). Default is \code{FALSE}, generating
unique groups combinations.}

\item{...}{additional parameters to pass to \code{\link[=calc_qvalues]{calc_qvalues()}}.}
}
\value{
tibble sorted by \code{q.value} of the test.
}
\description{
Data is assumed to be tidy with a single grouping variable; all other
variables are treated as activities to compare. Descriptive columns like
\code{cell_id} should not be included.
}
\details{
Applies \code{\link[stats:wilcox.test]{stats::wilcox.test()}} to unique pairs of groups for each measured
variable.
}
\examples{
x <- fsce_tidy[c("k_cluster", "Uracil_45", "riboG_44")]
stat_activity_grouped(x, group = k_cluster)

}
\references{
\doi{10.1038/nmeth.4612}
}
\concept{statistical tests}
