% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stat_anova_tukey}
\alias{stat_anova_tukey}
\title{Post-hoc analysis of ANOVA results}
\usage{
stat_anova_tukey(tbl, group = NULL, tidy = FALSE)
}
\arguments{
\item{group}{variable for generating combinations}

\item{tidy}{tidy the results}
}
\description{
Post-hoc analysis of ANOVA results
}
\examples{
x <- fsce_tidy[c("k_cluster", "Uracil_45", "riboG_44")]
x$k_cluster <- as.factor(x$k_cluster)

res <- stat_anova_grouped(x, k_cluster)
stat_anova_tukey(res, k_cluster, tidy = TRUE)

}
