% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy_coldata}
\alias{tidy_coldata}
\title{Tidy colData}
\usage{
tidy_coldata(fsce)
}
\arguments{
\item{fsce}{An object of class \code{\link{FunctionalSingleCellExperiment}}.}
}
\description{
Tidy colData
}
\section{Tidying}{

This method is a data tidier for a \code{\link{FunctionalSingleCellExperient}}. Returns
data from an \code{fsce} in a tidy format, where variables are columns and
observations are rows. Output will contain \code{experiment} and \code{cell_id} columns
that describe the source of the data.
}

\examples{
tidy_coldata(fsce_small)

}
\seealso{
Other tidiers: \code{\link{tidy_counts}},
  \code{\link{tidy_dims}}, \code{\link{tidy_logcounts}}
}
\concept{tidiers}
