% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy_coldata}
\alias{tidy_coldata}
\title{Tidy colData}
\usage{
tidy_coldata(sce)
}
\arguments{
\item{sce}{An object of class \code{\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment::SingleCellExperiment}}.}
}
\description{
Tidy colData
}
\section{Tidying}{

This is a data tidier for a \code{SingleCellExperient}. Returns
data from an \code{sce} in a tidy format, where variables are columns and
observations are rows.
}

\examples{
tidy_coldata(fsce_small)

}
\seealso{
Other tidiers: \code{\link{tidy_counts}},
  \code{\link{tidy_dims}}, \code{\link{tidy_logcounts}}
}
\concept{tidiers}
