% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy_dims}
\alias{tidy_dims}
\title{Tidy reducedDims data}
\usage{
tidy_dims(fsce, dimnames = NULL, dims = c(1, 2))
}
\arguments{
\item{fsce}{A \code{\link{FunctionalSingleCellExperiment}}.}

\item{dimnames}{vector of dimnames to retrieve. If \code{NULL}, retrieve all
dimnames.}

\item{dims}{vector of dimensions to retrieve}
}
\description{
Tidy reducedDims data
}
\section{Tidying}{

This method is a data tidier for a \code{\link{FunctionalSingleCellExperient}}. Returns
data from an \code{fsce} in a tidy format, where variables are columns and
observations are rows. Output will contain \code{experiment} and \code{cell_id} columns
that describe the source of the data.
}

