% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy_logcounts}
\alias{tidy_logcounts}
\title{Tidy logcounts data}
\usage{
tidy_logcounts(sce)
}
\arguments{
\item{sce}{An object of class \code{\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment::SingleCellExperiment}}.}
}
\description{
Tidy logcounts data
}
\section{Tidying}{

This is a data tidier for a \code{SingleCellExperient}. Returns
data from an \code{sce} in a tidy format, where variables are columns and
observations are rows.
}

\examples{
x <- fsce_small[ c("Uracil_45"), , "haircut"]
tidy_logcounts(x)
}
\seealso{
Other tidiers: \code{\link{tidy_coldata}},
  \code{\link{tidy_counts}}, \code{\link{tidy_dims}}
}
\concept{tidiers}
