% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{umitools_to_mtx}
\alias{umitools_to_mtx}
\title{Convert umitools flat format tsv to sparseMatrix .mtx format}
\usage{
umitools_to_mtx(count_file, output_path, ...)
}
\arguments{
\item{count_file}{path to umitools output file}

\item{output_path}{path for output files. matrix.mtx, barcodes.tsv and genes.tsv
will generated in supplied path}

\item{...}{additional arguments to pass to \link['readr:read_tsv()']{'readr::read_tsv()'}}
}
\description{
Convert umitools flat format tsv to sparseMatrix .mtx format
}
